+++ Controller/ImagineController.php
@@ -14,6 +14,7 @@ namespace Liip\ImagineBundle\Controller;
 use Imagine\Exception\RuntimeException;
 use Liip\ImagineBundle\Exception\Binary\Loader\NotLoadableException;
 use Liip\ImagineBundle\Exception\Imagine\Filter\NonExistingFilterException;
+use Liip\ImagineBundle\Imagine\Cache\Helper\PathHelper;
 use Liip\ImagineBundle\Imagine\Cache\SignerInterface;
 use Liip\ImagineBundle\Imagine\Data\DataManager;
 use Liip\ImagineBundle\Service\FilterService;
@@ -69,7 +70,7 @@ class ImagineController
      */
     public function filterAction(Request $request, $path, $filter)
     {
-        $path = urldecode($path);
+        $path = PathHelper::urlPathToFilePath($path);
         $resolver = $request->get('resolver');
 
         try {